+++ plugins/modules/oci_monitoring_alarm_facts.py
@@ -173,6 +173,14 @@ alarms:
             returned: on success
             type: str
             sample: body_example
+        is_notifications_per_metric_dimension_enabled:
+            description:
+                - "When set to `true`, splits notifications per metric stream. When set to `false`, groups notifications across metric streams.
+                  Example: `true`"
+                - Returned for get operation
+            returned: on success
+            type: bool
+            sample: true
         message_format:
             description:
                 - "The format to use for notification messages sent from this alarm. The formats are:
@@ -351,6 +359,7 @@ alarms:
         "resolution": "resolution_example",
         "pending_duration": "pending_duration_example",
         "body": "body_example",
+        "is_notifications_per_metric_dimension_enabled": true,
         "message_format": "RAW",
         "repeat_notification_duration": "repeat_notification_duration_example",
         "time_created": "2013-10-20T19:20:30+01:00",