+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [3.4.0]
+
+## Added
+
+- Support for preferred credentials for performing privileged operations in the Database Management service
+- Support for passing a content encoding when posting metrics in the Monitoring service
+
+## Fixed
+
+- Update migration guide for "oci_database_db_home_patch_history_facts". Fixes (#176)
+## Changed
+- Minimum required [OCI Python SDK](https://github.com/oracle/oci-python-sdk) changed to 2.84.0
+
 ## [3.3.0]
 
 ## Added